// src/api/errors/error-code.ts
// TODO: Add domain-specific error codes for your project below the general ones
export enum ErrorCode {
  // General errors
  INTERNAL_SERVER_ERROR = 'INTERNAL_SERVER_ERROR',
  BAD_REQUEST = 'BAD_REQUEST',
  UNAUTHORIZED = 'UNAUTHORIZED',
  FORBIDDEN = 'FORBIDDEN',
  NOT_FOUND = 'NOT_FOUND',
  CONFLICT = 'CONFLICT',

  // Auth errors
  AUTH_INVALID_REQUESTING_USER_DATA = 'INVALID_REQUESTING_USER_DATA',
  AUTH_ACCESS_PERMISSION_DENIED = 'ACCESS_PERMISSION_DENIED',

  // Messaging errors
  MESSAGE_PUBLISH_FAILED = 'MESSAGE_PUBLISH_FAILED',
  MESSAGE_INVALID_PARAMS = 'MESSAGE_INVALID_PARAMS',

  // Request validation errors
  INVALID_REQUEST_DATA = 'INVALID_REQUEST_DATA'
}